moved jaspify and qjit to the evaluation tools folder
