+++ src/qrisp/jasp/evaluation_tools/__init__.py
@@ -18,3 +18,5 @@
 
 from qrisp.jasp.evaluation_tools.boolean_simulation import *
 from qrisp.jasp.evaluation_tools.terminal_sampling import *
+from qrisp.jasp.evaluation_tools.jaspification import *
+from qrisp.jasp.evaluation_tools.catalyst_qjit import *
